backend/requirements.txt
fastapi[all]
sqlalchemy
passlib[bcrypt]
bcrypt==4.0.1
python-multipart
python-jose[cryptography]
psycopg2-binary
python-dotenv
uvicorn
pytest
pytest-cov
